Refactored solver
